requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
openai-whisper==20240930
python-multipart==0.0.5
